Cast CacheEntryIds to ints to get value comparison. Fixes bug where the
2008-07-20  Gabriel Burt  <gabriel.burt@gmail.com>

	* src/Core/Banshee.Services/Banshee.Collection.Database/DatabaseTrackInfo.cs:
	Cast CacheEntryIds to ints to get value comparison.  Fixes bug where the
	playback indicator next to the playing track would disappear, which was
	introduced when I committed the genre filter support and made the
	CacheEntryId an object.

svn path=/trunk/banshee/; revision=4221

+++ src/Core/Banshee.Services/Banshee.Collection.Database/DatabaseTrackInfo.cs
@@ -106,7 +106,6 @@ namespace Banshee.Collection.Database
             if (db_track == null) {
                 return base.TrackEqual (track);
             }
-            
             return TrackEqual (this, db_track);
         }
         
@@ -115,7 +114,7 @@ namespace Banshee.Collection.Database
             return a != null && b != null && 
                 a.TrackId == b.TrackId && 
                 a.CacheModelId == b.CacheModelId && 
-                a.CacheEntryId == b.CacheEntryId;
+                (int)a.CacheEntryId == (int)b.CacheEntryId;
         }
         
         public DatabaseArtistInfo Artist {
